scalar
