+++ datafusion/physical-expr-common/src/lib.rs
@@ -17,9 +17,12 @@
 
 pub mod aggregate;
 pub mod expressions;
+pub mod functions;
 pub mod intervals;
 pub mod physical_expr;
+pub mod scalar_function;
 pub mod sort_expr;
 pub mod sort_properties;
 pub mod tree_node;
+pub mod udf;
 pub mod utils;
