+++ datafusion/physical-expr/src/equivalence/properties.rs
@@ -1297,10 +1297,11 @@ mod tests {
     use std::ops::Not;
 
     use arrow::datatypes::{DataType, Field, Schema};
-    use arrow_schema::{Fields, TimeUnit};
+    use arrow_schema::{Fields, SortOptions, TimeUnit};
 
     use datafusion_common::DFSchema;
     use datafusion_expr::{Operator, ScalarUDF};
+    use datafusion_physical_expr_common::udf;
 
     use crate::equivalence::add_offset_to_expr;
     use crate::equivalence::tests::{
@@ -1791,7 +1792,7 @@ mod tests {
                 generate_table_for_eq_properties(&eq_properties, N_ELEMENTS, N_DISTINCT)?;
 
             let test_fun = ScalarUDF::new_from_impl(TestScalarUDF::new());
-            let floor_a = crate::udf::create_physical_expr(
+            let floor_a = udf::create_physical_expr(
                 &test_fun,
                 &[col("a", &test_schema)?],
                 &test_schema,
